Phase 1 Testing codes/requirements.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.23.5
mysql-connector-python==8.0.33